feat: add staking and update model

+++ contracts/src/events/started.cairo
@@ -2,7 +2,7 @@ pub use crate::events::index::Started;
 
 #[generate_trait]
 pub impl StartedImpl of StartedTrait {
-    fn new(player_id: felt252, game_id: u64, multiplier: u8) -> Started {
+    fn new(player_id: felt252, game_id: u64, multiplier: u16) -> Started {
         Started {
             player_id: player_id,
             game_id: game_id,
